Restructuring more things
