+++ packages/terra-framework-site/src/Index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './app/App';
-import { routes, navigation } from './app/appConfig';
+import { routes, navigation } from './appConfig';
 
 ReactDOM.render((
   <App routes={routes} navigation={navigation} />
