+++ packages/terra-framework-site/src/app/App.scss
@@ -9,3 +9,7 @@ body {
   position: relative;
   width: 100%;
 }
+
+.base {
+  height: 100%;
+}
