+++ packages/terra-framework-site/src/app/ApplicationMenu.jsx
@@ -1,10 +1,12 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import RoutingDelegate from 'terra-navigation/lib/routing/RoutingDelegate';
 
 import MenuList from './common/menu/MenuList';
 
 const propTypes = {
+  navigation: PropTypes.object,
   routingManager: RoutingDelegate.propType,
 };
