+++ packages/terra-framework-site/src/app/components/Components.jsx
@@ -9,6 +9,26 @@ const propTypes = {
   }),
 };
 
+const generateRoutes = config => (
+  Object.keys(config).map((componentKey) => {
+    const example = config[componentKey].example;
+    if (example) {
+      return (
+        <Route key={example.path} path={example.path} component={example.component} />
+      );
+    }
+    return undefined;
+  })
+);
+
+const generateRedirect = (config) => {
+  const firstExample = Object.keys(config).map(componentKey => (config[componentKey].example)).filter(example => !!example)[0];
+  if (firstExample) {
+    return <Redirect to={firstExample.path} />;
+  }
+  return null;
+};
+
 class Components extends React.Component {
   componentDidUpdate(prevProps) {
     if (this.element && prevProps.location.pathname !== this.props.location.pathname) {
@@ -22,22 +42,8 @@ class Components extends React.Component {
     return (
       <div ref={(element) => { this.element = element; }} style={{ height: '100%', position: 'relative', padding: '15px', overflow: 'auto' }}>
         <Switch>
-          {Object.keys(config).map((componentKey) => {
-            const example = config[componentKey].example;
-            if (example) {
-              return (
-                <Route key={example.path} path={example.path} component={example.component} />
-              );
-            }
-            return undefined;
-          })}
-          {(() => {
-            const firstExample = Object.keys(config).map(componentKey => (config[componentKey].example)).filter(example => !!example)[0];
-            if (firstExample) {
-              return <Redirect to={firstExample.path} />;
-            }
-            return null;
-          })()}
+          {generateRoutes(config)}
+          {generateRedirect(config)}
         </Switch>
       </div>
     );
