--- packages/terra-framework-site/src/app/tests/testConfig.js
@@ -1,14 +1,0 @@
-import LayoutTests from './LayoutTests';
-
-const menuConfig = {
-  '/tests/layout': {
-    path: '/tests/layout',
-    component: {
-      default: {
-        componentClass: LayoutTests,
-      },
-    },
-  },
-};
-
-export default menuConfig;
