+++ packages/terra-navigation/src/NavigationLayoutContent.scss
@@ -1,0 +1,5 @@
+:local {
+  .content {
+    height: 100%;
+  }
+}
